Added add item, up navigation (glitch), and main page improvements
The add item up navigation causes a crash at this point I am unsure how
to fix it. I placed add store in its own activity, and added a refresh
button to the main page, but more research is required to ensure the
data is substantiated after leaving the screen.